+++ monitoring/node-monitoring.yml
@@ -27,7 +27,7 @@ services:
     environment:
       METRICS_EXPORTER_BASIC_AUTH: "${BASIC_AUTH_USER_CREDENTIALS}"
       METRICS_EXPORTER_ROOT_API_URL: "${NODE_END_POINT}"
-      METRICS_EXPORTER_LEDGER_FOLDER: /data
+      #METRICS_EXPORTER_LEDGER_FOLDER: /data
       METRICS_EXPORTER_COLLECT_JSON_RPC_METRICS: "false"
 
   grafana:
